requirements.txt
langchain
openai
faiss-cpu
streamlit
python-dotenv
flask
tiktoken